move all screens to routes directory
