+++ composeApp/src/commonMain/kotlin/com/github/csc3380fall2024/team16/ui/routes/root/authenticated/home/bodybuilding/BodybuildingRoute.kt
@@ -2,7 +2,6 @@ package com.github.csc3380fall2024.team16.ui.routes.root.authenticated.home.body
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
-import com.github.csc3380fall2024.team16.ui.pages.home.BodybuildingPage
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -10,5 +9,5 @@ object BodybuildingRoute
 
 @Composable
 fun BodybuildingRoute.compose(navController: NavController) {
-    BodybuildingPage(navController)
+    BodybuildingScreen(navController)
 }
