+++ composeApp/src/commonMain/kotlin/com/github/csc3380fall2024/team16/ui/routes/root/authenticated/home/powerlifting/PowerliftingRoute.kt
@@ -2,7 +2,6 @@ package com.github.csc3380fall2024.team16.ui.routes.root.authenticated.home.powe
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
-import com.github.csc3380fall2024.team16.ui.pages.home.PowerliftingPage
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -10,5 +9,5 @@ object PowerliftingRoute
 
 @Composable
 fun PowerliftingRoute.compose(navController: NavController) {
-    PowerliftingPage(navController)
+    PowerliftingScreen(navController)
 }
