+++ composeApp/src/commonMain/kotlin/com/github/csc3380fall2024/team16/ui/routes/root/authenticated/home/weightloss/WeightLossRoute.kt
@@ -2,7 +2,6 @@ package com.github.csc3380fall2024.team16.ui.routes.root.authenticated.home.weig
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
-import com.github.csc3380fall2024.team16.ui.pages.home.WeightLossPage
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -10,5 +9,5 @@ object WeightLossRoute
 
 @Composable
 fun WeightLossRoute.compose(navController: NavController) {
-    WeightLossPage(navController)
+    WeightLossScreen(navController)
 }
