+++ composeApp/src/commonMain/kotlin/com/github/csc3380fall2024/team16/ui/routes/root/authenticated/news/NewsRoute.kt
@@ -2,7 +2,6 @@ package com.github.csc3380fall2024.team16.ui.routes.root.authenticated.news
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
-import com.github.csc3380fall2024.team16.ui.pages.NewsPage
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -10,5 +9,5 @@ object NewsRoute
 
 @Composable
 fun NewsRoute.compose(navController: NavController) {
-    NewsPage(navController)
+    NewsScreen(navController)
 }
