+++ composeApp/src/commonMain/kotlin/com/github/csc3380fall2024/team16/ui/routes/root/authenticated/social/SocialRoute.kt
@@ -2,7 +2,6 @@ package com.github.csc3380fall2024.team16.ui.routes.root.authenticated.social
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
-import com.github.csc3380fall2024.team16.ui.pages.SocialPage
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -10,5 +9,5 @@ object SocialRoute
 
 @Composable
 fun SocialRoute.compose(navController: NavController) {
-    SocialPage(navController)
+    SocialScreen(navController)
 }
