+++ composeApp/src/commonMain/kotlin/com/github/csc3380fall2024/team16/ui/routes/root/authenticated/tracker/TrackerRoute.kt
@@ -2,7 +2,6 @@ package com.github.csc3380fall2024.team16.ui.routes.root.authenticated.tracker
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
-import com.github.csc3380fall2024.team16.ui.pages.TrackerPage
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -10,5 +9,5 @@ object TrackerRoute
 
 @Composable
 fun TrackerRoute.compose(navController: NavController, currentCalories: Int, calorieGoal: Int) {
-    TrackerPage(navController, currentCalories, calorieGoal)
+    TrackerScreen(navController, currentCalories, calorieGoal)
 }
