+++ composeApp/src/commonMain/kotlin/com/github/csc3380fall2024/team16/ui/routes/root/unauthenticated/forgotPassword/ForgotPasswordRoute.kt
@@ -2,7 +2,6 @@ package com.github.csc3380fall2024.team16.ui.routes.root.unauthenticated.forgotP
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
-import com.github.csc3380fall2024.team16.ui.pages.ForgotPasswordPage
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -10,5 +9,5 @@ object ForgotPasswordRoute
 
 @Composable
 fun ForgotPasswordRoute.compose(navController: NavController) {
-    ForgotPasswordPage(navController)
+    ForgotPasswordScreen(navController)
 }
