+++ composeApp/src/commonMain/kotlin/com/github/csc3380fall2024/team16/ui/routes/root/unauthenticated/login/LoginScreen.kt
@@ -1,4 +1,4 @@
-package com.github.csc3380fall2024.team16.ui.pages
+package com.github.csc3380fall2024.team16.ui.routes.root.unauthenticated.login
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
@@ -22,47 +22,66 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.navigation.NavController
-import com.github.csc3380fall2024.team16.ui.routes.root.unauthenticated.login.LoginRoute
-import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
-@Preview
-fun ForgotPasswordPage(navController: NavController) {
+fun LoginScreen(
+    onLogin: (usernameOrEmail: String, password: String) -> Unit,
+    onNavigateRegister: () -> Unit,
+    onNavigateForgotPassword: () -> Unit,
+) {
+    var username by remember { mutableStateOf("") }
+    var password by remember { mutableStateOf("") }
+    
     Column(
         Modifier.fillMaxSize().padding(horizontal = 20.dp),
         horizontalAlignment = Alignment.CenterHorizontally,
         verticalArrangement = Arrangement.spacedBy(40.dp, Alignment.CenterVertically),
     ) {
-        Text("Forgot Password", fontSize = 30.sp, fontWeight = FontWeight.Bold)
+        Text("Login", fontSize = 30.sp, fontWeight = FontWeight.Bold)
+        Text("Welcome back to Universal Fitness.")
         
         Column(
             Modifier.padding(horizontal = 20.dp),
-            verticalArrangement = Arrangement.spacedBy(
-                20.dp,
-                Alignment.CenterVertically
-            )
+            verticalArrangement = Arrangement.spacedBy(20.dp, Alignment.CenterVertically)
         ) {
-            var username by remember { mutableStateOf("") }
             TextField(
                 username,
                 { username = it },
                 Modifier.fillMaxWidth(),
-                placeholder = { Text("Enter Email") },
+                placeholder = { Text("Username") },
                 keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Email),
+                singleLine = true
+            )
+            
+            TextField(
+                password,
+                { password = it },
+                Modifier.fillMaxWidth(),
+                placeholder = { Text("Password") },
+                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password),
                 singleLine = true
             )
         }
         
-        Button(onClick = {}, Modifier.fillMaxWidth().padding(20.dp)) {
-            Text("Submit", fontSize = 20.sp, lineHeight = 5.sp)
+        Button({
+            onLogin(username, password)
+        }, Modifier.fillMaxWidth().padding(20.dp)) {
+            Text("Login", fontSize = 20.sp, lineHeight = 5.sp)
         }
         
         Text(
-            text = "Back to Login",
+            text = "Register",
             style = MaterialTheme.typography.titleLarge,
             modifier = Modifier.clickable {
-                navController.navigate(LoginRoute)
+                onNavigateRegister()
+            }
+        )
+        
+        Text(
+            text = "Forgot Password",
+            style = MaterialTheme.typography.titleLarge,
+            modifier = Modifier.clickable {
+                onNavigateForgotPassword()
             }
         )
     }
