+++ composeApp/src/commonMain/kotlin/com/github/csc3380fall2024/team16/ui/routes/root/unauthenticated/register/RegisterRoute.kt
@@ -4,7 +4,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.github.csc3380fall2024.team16.RpcClient
-import com.github.csc3380fall2024.team16.ui.pages.RegisterPage
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -13,7 +12,7 @@ object RegisterRoute
 @Composable
 fun RegisterRoute.compose(client: RpcClient, onAuthenticated: (String) -> Unit) {
     val viewModel = viewModel { RegisterViewModel(client) }
-    RegisterPage(onRegister = viewModel::register)
+    RegisterScreen(onRegister = viewModel::register)
     LaunchedEffect(viewModel.token) {
         val token = viewModel.token
         if (token != null) onAuthenticated(token)
